gui 작업을 위해 폴더 개편 + gui setting_시작화면
