+++ Codes/subgroup3/UC_09/UC_09_ShopkeeperOrder/DBconnection.py
@@ -1,4 +1,4 @@
-import OrderList
+from . import OrderList
 
 class DBconnection:
     def __init__(self, result, order_list):
