+++ Codes/subgroup3/UC_09/UC_09_ShopkeeperOrder/OrderAccepter.py
@@ -1,4 +1,4 @@
-import OrderList
+from . import OrderList
 import os
 
 class OrderAccepter:
