+++ Codes/subgroup3/UC_09/UC_09_ShopkeeperOrder/OrderListMaker.py
@@ -1,6 +1,6 @@
 import datetime as dt
-import OrderSender as ods
-import OrderList as ol
+from . import OrderSender as ods
+from . import OrderList as ol
 
 class OrderListMaker:
     def __init__(self, connected_wh):
