+++ Codes/subgroup3/UC_09/창고주.txt
@@ -1,0 +1,4 @@
+id : w20210527
+connected_s : s20210527
+money : 2000000
+Auto_order_mode : off
